Src fix (DB version was off)
IMPORTANT: Remember about doing a backup of the database right before the COMMIT and PUSH in order to save DB changes (regardless of whether you manually changed something in the DB), and doing a restore right after the PULL in order to have a correct database version as it is NOT directly versioned (only a dump is). If DB scripts do not work, try chmod. 'install' directory removed after installation.